Implementing cache.
