+++ lib/app_creator.js
@@ -167,6 +167,7 @@ module.exports = function(appRouter, path){
 
 	// create routes
 	var routers = routeManager(app);
+	var cacheUpdate = clientCache(app);
 
 	// http req router
 	httpRouter(app);
@@ -182,24 +183,21 @@ module.exports = function(appRouter, path){
 	app.start = function(cb){
 		var pid = new Date().getTime() + Math.random();
 		startProcessId = pid;
-		clientCache.create(app, pid, routers['client'].list(), function(){
-			if(startProcessId !== pid) {
-				if(startProcessId) clientCache.remove(app, pid);
-				else clientCache.apply(app, pid);
-				return;
-			}
+		console.log('Starting app: ' + app.config.app.title);
+		cacheUpdate(app, routers['client'].list(), function(){
+			if(startProcessId !== pid) return;
+			if(fw.debug) console.log('Connecting to database: ' + app.config.db.type);
 			initDb(function(){
 				if(startProcessId !== pid) {
-					if(startProcessId) clientCache.remove(app, pid);
-					else clientCache.apply(app, pid);
 					(closeDb || closeDbDefault)(cb);
 					return;
 				}
-				clientCache.apply(app, pid);
+				// TODO load server modules
 				app.enabled = true;
 				for(var i=0; i<app.config.app.host.length; i++) {
 					appRouter.bindHost(app, app.config.app.host[i]);
 				}
+				console.log('App started: ' + app.config.app.title);
 				cb();
 			});
 		});
