+++ lib/routers/module.js
@@ -8,8 +8,6 @@ module.exports = function(app, prefix, dir, cb){
 	var routes = {};
 	var tmplManager = app.route.tmpl.addDir(dirType, prefix, path);
 
-	// TODO load modules
-
 	// listen files
 	var listener = fsListener(dir, function(err, watcher){
 		if(err) {
