+++ lib/utils.js
@@ -47,18 +47,27 @@ var walkFileTree = exports.walkFileTree = function(root, cb, doneCb, curPath){
 };
 
 // rmdirp
-exports.rmdirp = function(path) {
-	var files = [];
-	if(fs.existsSync(path)) {
-		files = fs.readdirSync(path);
-		files.forEach(function(file,index){
-			var curPath = path + '/' + file;
-			if(fs.statSync(curPath).isDirectory()) {
-				rmdirp(curPath);
-			} else {
-				fs.unlinkSync(curPath);
-			}
-		});
-		fs.rmdirSync(path);
-	}
+var rmdirp = exports.rmdirp = function(root, cb) {
+	fs.stat(root, function(err, stat){
+		if(err) return cb();
+		if(stat.isDirectory()) {
+			fs.readdir(function(err, files){
+				if(err) return cb();
+				var finishCount = files.length + 1;
+				var finish = function(){
+					finishCount--;
+					if(!finishCount) {
+						fs.rmdir(root, cb);
+					}
+				};
+				files.forEach(function(file){
+					file = root + '/' + file;
+					rmdirp(file, finish);
+				});
+				finish();
+			});
+		} else {
+			fs.unlink(root, cb);
+		}
+	});
 };
